feat: user summary moved to a component
